ver 4.21.01 updating to Qt5

+++ libpp/src/TokenizerIterator.h
@@ -1,7 +1,7 @@
 #ifndef TokenizerIterator_H
 #define TokenizerIterator_H
 
-#include <string>
+#include <QString>
 #include "Token.h"
 
 class Tokenizer;
@@ -9,17 +9,17 @@ class Tokenizer;
 class TokenizerIterator
 {
 public:
-	TokenizerIterator(const Tokenizer* tokenizer, size_t index = 0);
-	TokenizerIterator(Tokenizer* tokenizer, size_t index = 0);
-	~TokenizerIterator() {}
-	const Token& operator*() const;
-	Token& operator*();
-	TokenizerIterator& operator++();
-	bool operator==(const TokenizerIterator& rhs) const;
-	bool operator!=(const TokenizerIterator& rhs) const;
+    TokenizerIterator(const Tokenizer* tokenizer, int index = 0);
+    TokenizerIterator(Tokenizer* tokenizer, int index = 0);
+    ~TokenizerIterator() {}
+    const Token& operator*() const;
+    Token& operator*();
+    TokenizerIterator& operator++();
+    bool operator==(const TokenizerIterator& rhs) const;
+    bool operator!=(const TokenizerIterator& rhs) const;
 private:
-	const Tokenizer* const_tokenizer_;
-	Tokenizer* tokenizer_;
-	size_t currentIndex_;
+    const Tokenizer* const_tokenizer_;
+    Tokenizer* tokenizer_;
+    int currentIndex_;
 };
 #endif
